- functional equations of motion - functional propulsion - functional lon aero - python plotting script - still need to sniff out some bugs making it hard to trim

+++ src/Simulation.h
@@ -17,7 +17,7 @@ class Simulation{
 	public:
 		Simulation();
 		void printTimestep();
-		void run(float maxTime);
+		void run(double maxTime);
 		void toCSV(const std::string& filename);
 };
